+++ src/CodeGen_Internal.h
@@ -55,6 +55,15 @@ Expr lower_euclidean_div(Expr a, Expr b);
 Expr lower_euclidean_mod(Expr a, Expr b);
 ///@}
 
+/** Given an llvm::Module, set llvm:TargetOptions, cpu and attr information */
+void get_target_options(const llvm::Module &module, llvm::TargetOptions &options, std::string &mcpu, std::string &mattrs);
+
+/** Given two llvm::Modules, clone target options from one to the other */
+void clone_target_options(const llvm::Module &from, llvm::Module &to);
+
+/** Given an llvm::Module, get or create an llvm:TargetMachine */
+llvm::TargetMachine *get_target_machine(const llvm::Module &module);
+
 }}
 
 #endif
